Reorganize how the --show option is handled
1. Keep track of which options have been used.
2. No longer run the pipeline show options multiple times
   for pipetask qgraph.
3. No longer warn about options not being used when they
   are about to be used.
4. If --show is used, no longer continue on unless there are
   remaining --show options to process.

The latter means that using `pipetask qgraph` with only pipeline
show options exits without building the graph, and any use of
--show leads to `pipetask run` not running anything.
